+++ src/mini/Vulkan/Ctors.hpp
@@ -33,7 +33,7 @@ namespace mini::vk
         std::ifstream file(path, std::ios::ate | std::ios::binary);
         mini::Assert(file.is_open(), "cannot open shader file");
 
-        const uint32_t size = file.tellg();
+        const uint32_t size = (uint32_t)file.tellg();
         auto ptrBuffer = mini::mem::ClaimBlock<char[10000]>(); //! capacity
         //char buffer[10000]; //!be aware of capacity, maybe use allocator to not exhaust stack
         file.seekg(std::ios::beg);
